+++ src/app.module.ts
@@ -7,7 +7,7 @@ import { RegisterModule } from './register/register.module';
 import { LoginModule } from './login/login.module';
 import { ForgetPasswordModule } from './forgetPassword/forgetPassword.module';
 import { UsersModule } from './users/users.module';
-import { PredictionModule } from './users/prediction/prediction.module';
+import { PredictionModule } from './users/predictions/predictions.module';
 
 @Module({
   imports: [
