+++ src/users/predictions/predictions.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { PredictionController } from './prediction.controller';
-import { PredictionService } from './prediction.service';
+import { PredictionController } from './predictions.controller';
+import { PredictionService } from './predictions.service';
 @Module({
   controllers: [PredictionController],
   providers: [PredictionService],
